Model/requirements.txt
opencv-contrib-python
numpy
matplotlib
tensorflow
keras
python-socketio
Flask
eventlet
opencv-python
websockets
Pillow
opencv-python>=4.5.4